refactor(js&sign): 调整部分代码
