+++ signature/gen/sign_mods.go
@@ -1,16 +1,11 @@
 package main
 
 import (
-	"crypto"
-	"crypto/rand"
-	"crypto/rsa"
-	"crypto/x509"
-	"encoding/base64"
-	"encoding/pem"
 	"fmt"
 	"log"
 	"os"
 	"path/filepath"
+	sealcrypto "sealdice-core/utils/crypto"
 	"strings"
 )
 
@@ -52,46 +47,14 @@ func signModFile(privateKey string, path, name string, signStrTmpl string) {
 		return
 	}
 	data, _ := os.ReadFile(path)
-	sign, err := rsaSign(data, privateKey)
+	sign, err := sealcrypto.RSASign(data, privateKey)
 	if err != nil {
 		return
 	}
-	newData := []byte(fmt.Sprintf(signStrTmpl, sign))
+	newData := []byte(fmt.Sprintf(signStrTmpl, sign) + string(data))
 	err = os.WriteFile("[signed]"+name, newData, 0644)
 	if err != nil {
 		return
 	}
 	log.Printf("完成对 %s 的签名，签名后的文件为 [signed]%s\n", name, name)
-}
-
-// rsaSign 签名
-func rsaSign(data []byte, privateKey string) (string, error) {
-	key := readPrivateKey(privateKey)
-	hashed := calculateSHA512(data)
-	sign, err := rsa.SignPSS(rand.Reader, key, crypto.SHA512, hashed, nil)
-	if err != nil {
-		return "", err
-	}
-	return base64.StdEncoding.EncodeToString(sign), nil
-}
-
-func calculateSHA512(data []byte) []byte {
-	hashInstance := crypto.SHA512.New()
-	hashInstance.Write(data)
-	hashed := hashInstance.Sum(nil)
-	return hashed
-}
-
-// readPrivateKey 读取私钥
-func readPrivateKey(privateKey string) *rsa.PrivateKey {
-	block, _ := pem.Decode([]byte(privateKey))
-	if block == nil {
-		return nil
-	}
-	privateKeyInterface, err := x509.ParsePKCS8PrivateKey(block.Bytes)
-	if err != nil {
-		return nil
-	}
-	key := privateKeyInterface.(*rsa.PrivateKey)
-	return key
 }
